Commonから文字列操作に関するメソッドを分離
